requirements.txt
python-dotenv
chardet